requirements.txt
PyYAML==6.0
pandas==1.5.3
requests==2.28.2
pytest==7.2.0
